refactor(api): rename get into find and use withTransaction
